+++ apps/frontend/src/app/business/lp/page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function BusinessLp() {
+    return (
+        <div>
+            <h2 className="flex justify-center text-2xl font-semibold">このページがLPですかね！</h2>
+        </div>
+    );
+}
