+++ apps/frontend/src/app/business/page.tsx
@@ -1,9 +1,26 @@
+"use client";
+
 import React from "react";
 
-export default function BusinessLp() {
+export default function Report() {
     return (
-        <div>
-            <h2 className="flex justify-center text-2xl font-semibold">このページがLPですかね！</h2>
+        <div className="container flex justify-center mx-auto p-3">
+            <div className="w-96">
+                <form className="max-w-md mx-auto">
+                    <input
+                        type="search"
+                        id="search"
+                        className="block w-full p-2 text-sm text-gray-900 border border-gray-300 rounded-lg bg-gray-50 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                        placeholder="（例）リムジン"
+                    />
+                    <button
+                        type="submit"
+                        className="text-white absolute end-2.5 bottom-2.5 bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
+                    >
+                        検索
+                    </button>
+                </form>
+            </div>
         </div>
     );
 }
