+++ apps/frontend/src/app/commons/constants/constants.ts
@@ -7,7 +7,7 @@ export const CUSTOMER_NAVIGATION_LINKS = [
 
 // to business
 export const BUSINESS_NAVIGATION_LINKS = [
-    { label: "ホーム", href: "/business" },
-    { label: "レポート", href: "/business/report" },
+    { label: "ホーム", href: "/business/lp" },
+    { label: "レポート", href: "/business" },
     { label: "ログイン", href: "/business/login" },
 ];
